glsl: reduce number of if statements in IR due to switches
While upstream mesa has for a while implemented switch statements as
loops in its IR, we reverted this change due to slow shader
compilation times on some platforms caused by the loops. Instead, a
series of if statements is used. However, we also encounter issues on
some platforms due to there being too many if statements.

Currently for each case in a switch we:
1. OR-assign the fallthrough variable with the condition.
2. Conditionally set the fallthrough variable to false if the break
   variable is true.
3. Conditionally execute the case's statements if the fallthrough
   variable is true.

This results in 2 if statements per case in the output optimized GLSL
code. Instead of step number 2, this change makes it so that we
AND the fallthrough variable with !break_var. This logically has the
same effect, but requires half as many if statements which hopefully
will avoid driver bugs on various platforms.